Combine Rack::MockSession into Rack::Test::Session
This is basically the reverse of cd40d6191dbc406d44b10d96208a4254c96988da.
There doesn't seem to be any reason to split the class up and have two
separate sessions, especially when one depends so much on the other.

This makes Rack::MockSession an alias of Rack::Test::Session, and
makes Rack::Test::Session.new handle being called with a
Rack::MockSession (returning the argument in that case).

This also tries to keep backwards compatibility in
Rack::Test::Methods#build_rack_test_session by calling
build_rack_mock_session if it is defined.  At least Capybara uses
build_rack_mock_session.

Simplify Rack::Test::Methods#current_session by keeping a
variable for the current session.  Make it handle exceptions
better, restoring the previous state. Remove _current_session_names,
since there is no longer a reason to keep an array of session names.

This change passes all existing specs.  In spite of the backwards
compatibility, considering the changes, I plan to do a major bump
for the next release.
